add command to check a single safe
