+++ governance/package.json
@@ -32,7 +32,7 @@
     "test": "hardhat test",
     "ci": "yarn lint && yarn test",
     "check": "node ./all_networks unlock:info --quiet",
-    "check:multisig": "yarn hardhat run scripts/multisig/info.js",
+    "check:multisig": "yarn hardhat run scripts/multisig/check-all.js",
     "lint:contracts": "solhint 'contracts/**/*.sol'",
     "lint:code": "eslint --resolve-plugins-relative-to ../packages/eslint-config .",
     "lint": "yarn lint:contracts && yarn lint:code"
